[ops] move !fetch-url-* to ops/fetch_url.py
